add SimChat trait to eventually allow holochain-simchat and lib3h-simchat to use the same interface
